requirements.txt
flask
replicate
google-cloud-secret-manager
python-dotenv
openai